requirements.txt
Django>=4.2,<5.0
gunicorn
psycopg2-binary
dj-database-url
djangorestframework
djangorestframework-simplejwt==5.5.1
pytest
pytest-django
pytest-cov
